+++ lib/src/telegram/models/transaction_partner_user.dart
@@ -17,12 +17,20 @@ class TransactionPartnerUser extends TransactionPartner {
   /// Optional. Bot-specified paid media payload
   final String? paidMediaPayload;
 
+  /// Optional. The duration of the paid subscription.
+  final int? subscriptionPeriod;
+
+  /// Optional. The gift sent to the user by the bot.
+  final Gift? gift;
+
   /// Constructs a [TransactionPartnerUser] object.
   const TransactionPartnerUser({
     required this.user,
     this.invoicePayload,
     this.paidMedia,
     this.paidMediaPayload,
+    this.subscriptionPeriod,
+    this.gift,
   });
 
   /// Creates a [TransactionPartnerUser] object from JSON.
@@ -38,6 +46,8 @@ class TransactionPartnerUser extends TransactionPartner {
             )
           : null,
       paidMediaPayload: json['paid_media_payload'],
+      subscriptionPeriod: json['subscription_period'],
+      gift: json['gift'] != null ? Gift.fromJson(json['gift']) : null,
     );
   }
 
@@ -50,6 +60,8 @@ class TransactionPartnerUser extends TransactionPartner {
       'invoice_payload': invoicePayload,
       'paid_media': paidMedia?.map((e) => e.toJson()).toList(),
       'paid_media_payload': paidMediaPayload,
+      'subscription_period': subscriptionPeriod,
+      'gift': gift?.toJson(),
     }..removeWhere(_nullFilter);
   }
 }